requirements.txt
absl-py==0.9.0
appdirs==1.4.4
appnope==0.1.2
asgiref==3.2.10
astunparse==1.6.3
attrs==19.3.0
aws==0.2.5
awscli==1.18.97
awsebcli==3.18.1
backcall==0.2.0
bcrypt==3.1.7
blessed==1.17.8
boto==2.49.0
botocore==1.15.49
cached-property==1.5.1
cachetools==4.1.0
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cmake==3.20.4
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
distlib==0.3.1
Django==3.0.5
djongo==1.3.3
dlib==19.22.0
docker==4.2.2
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
et-xmlfile==1.0.1
fabric==2.5.0
filelock==3.0.12
future==0.16.0
gast==0.3.3
glob3==0.0.1
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.7
importlib-metadata==1.6.0
invoke==1.4.1
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.1
Markdown==3.2.2
markdown2==2.4.0
matplotlib==3.3.3
matplotlib-inline==0.1.2
MouseInfo==0.1.3
multitasking==0.0.9
mysql-connector-python==8.0.22
nltk==3.5
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.1.1.26
openpyxl==2.6.2
opt-einsum==3.2.1
pandas==1.0.4
pandas-datareader==0.8.1
paramiko==2.7.1
parso==0.8.2
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pipenv==2020.6.2
prettytable==0.7.2
prompt-toolkit==3.0.19
protobuf==3.12.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
PyAutoGUI==0.9.50
pycparser==2.20
pygame==1.9.6
PyGetWindow==0.0.8
Pygments==2.9.0
PyInputPlus==0.2.9
pymongo==3.10.1
PyMsgBox==1.0.8
PyNaCl==1.4.0
pynput==1.7.1
pyobjc==6.2.2
pyobjc-core==6.2.2
pyobjc-framework-Accounts==6.2.2
pyobjc-framework-AddressBook==6.2.2
pyobjc-framework-AdSupport==6.2.2
pyobjc-framework-AppleScriptKit==6.2.2
pyobjc-framework-AppleScriptObjC==6.2.2
pyobjc-framework-ApplicationServices==6.2.2
pyobjc-framework-AuthenticationServices==6.2.2
pyobjc-framework-AutomaticAssessmentConfiguration==6.2.2
pyobjc-framework-Automator==6.2.2
pyobjc-framework-AVFoundation==6.2.2
pyobjc-framework-AVKit==6.2.2
pyobjc-framework-BusinessChat==6.2.2
pyobjc-framework-CalendarStore==6.2.2
pyobjc-framework-CFNetwork==6.2.2
pyobjc-framework-CloudKit==6.2.2
pyobjc-framework-Cocoa==6.2.2
pyobjc-framework-Collaboration==6.2.2
pyobjc-framework-ColorSync==6.2.2
pyobjc-framework-Contacts==6.2.2
pyobjc-framework-ContactsUI==6.2.2
pyobjc-framework-CoreAudio==6.2.2
pyobjc-framework-CoreAudioKit==6.2.2
pyobjc-framework-CoreBluetooth==6.2.2
pyobjc-framework-CoreData==6.2.2
pyobjc-framework-CoreHaptics==6.2.2
pyobjc-framework-CoreLocation==6.2.2
pyobjc-framework-CoreMedia==6.2.2
pyobjc-framework-CoreMediaIO==6.2.2
pyobjc-framework-CoreML==6.2.2
pyobjc-framework-CoreMotion==6.2.2
pyobjc-framework-CoreServices==6.2.2
pyobjc-framework-CoreSpotlight==6.2.2
pyobjc-framework-CoreText==6.2.2
pyobjc-framework-CoreWLAN==6.2.2
pyobjc-framework-CryptoTokenKit==6.2.2
pyobjc-framework-DeviceCheck==6.2.2
pyobjc-framework-DictionaryServices==6.2.2
pyobjc-framework-DiscRecording==6.2.2
pyobjc-framework-DiscRecordingUI==6.2.2
pyobjc-framework-DiskArbitration==6.2.2
pyobjc-framework-DVDPlayback==6.2.2
pyobjc-framework-EventKit==6.2.2
pyobjc-framework-ExceptionHandling==6.2.2
pyobjc-framework-ExecutionPolicy==6.2.2
pyobjc-framework-ExternalAccessory==6.2.2
pyobjc-framework-FileProvider==6.2.2
pyobjc-framework-FileProviderUI==6.2.2
pyobjc-framework-FinderSync==6.2.2
pyobjc-framework-FSEvents==6.2.2
pyobjc-framework-GameCenter==6.2.2
pyobjc-framework-GameController==6.2.2
pyobjc-framework-GameKit==6.2.2
pyobjc-framework-GameplayKit==6.2.2
pyobjc-framework-ImageCaptureCore==6.2.2
pyobjc-framework-IMServicePlugIn==6.2.2
pyobjc-framework-InputMethodKit==6.2.2
pyobjc-framework-InstallerPlugins==6.2.2
pyobjc-framework-InstantMessage==6.2.2
pyobjc-framework-Intents==6.2.2
pyobjc-framework-IOSurface==6.2.2
pyobjc-framework-iTunesLibrary==6.2.2
pyobjc-framework-LatentSemanticMapping==6.2.2
pyobjc-framework-LaunchServices==6.2.2
pyobjc-framework-libdispatch==6.2.2
pyobjc-framework-LinkPresentation==6.2.2
pyobjc-framework-LocalAuthentication==6.2.2
pyobjc-framework-MapKit==6.2.2
pyobjc-framework-MediaAccessibility==6.2.2
pyobjc-framework-MediaLibrary==6.2.2
pyobjc-framework-MediaPlayer==6.2.2
pyobjc-framework-MediaToolbox==6.2.2
pyobjc-framework-Metal==6.2.2
pyobjc-framework-MetalKit==6.2.2
pyobjc-framework-ModelIO==6.2.2
pyobjc-framework-MultipeerConnectivity==6.2.2
pyobjc-framework-NaturalLanguage==6.2.2
pyobjc-framework-NetFS==6.2.2
pyobjc-framework-Network==6.2.2
pyobjc-framework-NetworkExtension==6.2.2
pyobjc-framework-NotificationCenter==6.2.2
pyobjc-framework-OpenDirectory==6.2.2
pyobjc-framework-OSAKit==6.2.2
pyobjc-framework-OSLog==6.2.2
pyobjc-framework-PencilKit==6.2.2
pyobjc-framework-Photos==6.2.2
pyobjc-framework-PhotosUI==6.2.2
pyobjc-framework-PreferencePanes==6.2.2
pyobjc-framework-PushKit==6.2.2
pyobjc-framework-Quartz==6.2.2
pyobjc-framework-QuickLookThumbnailing==6.2.2
pyobjc-framework-SafariServices==6.2.2
pyobjc-framework-SceneKit==6.2.2
pyobjc-framework-ScreenSaver==6.2.2
pyobjc-framework-ScriptingBridge==6.2.2
pyobjc-framework-SearchKit==6.2.2
pyobjc-framework-Security==6.2.2
pyobjc-framework-SecurityFoundation==6.2.2
pyobjc-framework-SecurityInterface==6.2.2
pyobjc-framework-ServiceManagement==6.2.2
pyobjc-framework-Social==6.2.2
pyobjc-framework-SoundAnalysis==6.2.2
pyobjc-framework-Speech==6.2.2
pyobjc-framework-SpriteKit==6.2.2
pyobjc-framework-StoreKit==6.2.2
pyobjc-framework-SyncServices==6.2.2
pyobjc-framework-SystemConfiguration==6.2.2
pyobjc-framework-SystemExtensions==6.2.2
pyobjc-framework-UserNotifications==6.2.2
pyobjc-framework-VideoSubscriberAccount==6.2.2
pyobjc-framework-VideoToolbox==6.2.2
pyobjc-framework-Vision==6.2.2
pyobjc-framework-WebKit==6.2.2
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
PySimpleValidate==0.2.10
python-dateutil==2.8.0
python-dotenv==0.19.0
PyTweening==1.0.3
pytz==2020.1
PyYAML==5.3.1
pyzmq==22.1.0
regex==2020.5.14
requests==2.20.1
requests-oauthlib==1.3.0
rsa==4.0
rubicon-objc==0.4.0
s3transfer==0.3.3
scikit-learn==0.23.0
scipy==1.4.1
selenium==3.141.0
semantic-version==2.5.0
six==1.11.0
sqlparse==0.2.4
stdiomask==0.0.5
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
texttable==1.6.2
threadpoolctl==2.0.0
tornado==6.1
tqdm==4.46.0
traitlets==5.0.5
urllib3==1.24.3
vidstream==0.0.10
virtualenv==20.0.25
virtualenv-clone==0.5.4
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
yfinance==0.1.54
zipp==3.1.0
